+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `Pledge.race` and make it so a Pledge's state can't change
 * Make `Pledge` class global (if allowed)
 * Add `String#startsWithAny` and `String#endsWithAny`
+* Add `Object.values` and `Object.entries` polyfill
 
 ## 0.5.5 (2018-06-18)